max borrow test updated.

+++ contracts/compound/Comptroller.sol
@@ -407,7 +407,7 @@ contract Comptroller is ComptrollerV3Storage, ComptrollerInterface, ComptrollerE
     // Get max borrow or redeem considering cToken liquidity
     uint256 cTokenLiquidity = cTokenModify.getCash();
 
-    uint maxAmount = (maxBorrowOrRedeemAmount * 10**cTokenModify.decimals()) / 1e18;
+    uint256 maxAmount = (maxBorrowOrRedeemAmount * 10**cTokenModify.decimals()) / 1e18;
 
     // Return the minimum of the two maximums
     return maxAmount <= cTokenLiquidity ? maxAmount : cTokenLiquidity;
